third_party/blink/web_tests/external/wpt/css/css-values/calc-size/animation/interpolate-size-min-height-interpolation-expected.txt

This is a testharness.js-based test.
Found 66 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (-0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (0) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (-0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (0) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (-0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (0) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (-0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (0) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (-0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (0) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (-0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (0) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [fit-content] to neutral at (-0.3) should be [calc-size(fit-content, -3px + size * 1.3)]
  assert_equals: expected "calc - size ( fit - content , - 3px + size * 1.3 ) " but got "calc - size ( fit - content , - 3px + 0px + size * 1.3 ) "
[FAIL] CSS Animations: property <min-height> from [fit-content] to neutral at (0) should be [calc-size(fit-content, 0px + size * 1)]
  assert_equals: expected "calc - size ( fit - content , 0px + size * 1 ) " but got "calc - size ( fit - content , 0px + 0px + size * 1 ) "
[FAIL] CSS Animations: property <min-height> from [fit-content] to neutral at (0.3) should be [calc-size(fit-content, 3px + size * 0.7)]
  assert_equals: expected "calc - size ( fit - content , 3px + size * 0.7 ) " but got "calc - size ( fit - content , 3px + 0px + size * 0.7 ) "
[FAIL] CSS Animations: property <min-height> from [fit-content] to neutral at (0.6) should be [calc-size(fit-content, 6px + size * 0.4)]
  assert_equals: expected "calc - size ( fit - content , 6px + size * 0.4 ) " but got "calc - size ( fit - content , 6px + 0px + size * 0.4 ) "
[FAIL] CSS Animations: property <min-height> from [fit-content] to neutral at (1) should be [calc-size(fit-content, 10px + size * 0)]
  assert_equals: expected "calc - size ( fit - content , 10px + size * 0 ) " but got "calc - size ( fit - content , 10px + 0px + size * 0 ) "
[FAIL] CSS Animations: property <min-height> from [fit-content] to neutral at (1.5) should be [calc-size(fit-content, 15px + size * -0.5)]
  assert_equals: expected "calc - size ( fit - content , 15px + size * - 0.5 ) " but got "calc - size ( fit - content , 15px + 0px + size * - 0.5 ) "
[FAIL] Web Animations: property <min-height> from [fit-content] to neutral at (-0.3) should be [calc-size(fit-content, -3px + size * 1.3)]
  assert_equals: expected "calc - size ( fit - content , - 3px + size * 1.3 ) " but got "calc - size ( fit - content , - 3px + 0px + size * 1.3 ) "
[FAIL] Web Animations: property <min-height> from [fit-content] to neutral at (0) should be [calc-size(fit-content, 0px + size * 1)]
  assert_equals: expected "calc - size ( fit - content , 0px + size * 1 ) " but got "calc - size ( fit - content , 0px + 0px + size * 1 ) "
[FAIL] Web Animations: property <min-height> from [fit-content] to neutral at (0.3) should be [calc-size(fit-content, 3px + size * 0.7)]
  assert_equals: expected "calc - size ( fit - content , 3px + size * 0.7 ) " but got "calc - size ( fit - content , 3px + 0px + size * 0.7 ) "
[FAIL] Web Animations: property <min-height> from [fit-content] to neutral at (0.6) should be [calc-size(fit-content, 6px + size * 0.4)]
  assert_equals: expected "calc - size ( fit - content , 6px + size * 0.4 ) " but got "calc - size ( fit - content , 6px + 0px + size * 0.4 ) "
[FAIL] Web Animations: property <min-height> from [fit-content] to neutral at (1) should be [calc-size(fit-content, 10px + size * 0)]
  assert_equals: expected "calc - size ( fit - content , 10px + size * 0 ) " but got "calc - size ( fit - content , 10px + 0px + size * 0 ) "
[FAIL] Web Animations: property <min-height> from [fit-content] to neutral at (1.5) should be [calc-size(fit-content, 15px + size * -0.5)]
  assert_equals: expected "calc - size ( fit - content , 15px + size * - 0.5 ) " but got "calc - size ( fit - content , 15px + 0px + size * - 0.5 ) "
[FAIL] CSS Animations: property <min-height> from neutral to [min-content] at (-0.3) should be [calc-size(min-content, 13px + size * -0.3)]
  assert_equals: expected "calc - size ( min - content , 13px + size * - 0.3 ) " but got "calc - size ( min - content , 13px + 0px + size * - 0.3 ) "
[FAIL] CSS Animations: property <min-height> from neutral to [min-content] at (0) should be [calc-size(min-content, 10px + size * 0)]
  assert_equals: expected "calc - size ( min - content , 10px + size * 0 ) " but got "calc - size ( min - content , 10px + 0px + size * 0 ) "
[FAIL] CSS Animations: property <min-height> from neutral to [min-content] at (0.3) should be [calc-size(min-content, 7px + size * 0.3)]
  assert_equals: expected "calc - size ( min - content , 7px + size * 0.3 ) " but got "calc - size ( min - content , 7px + 0px + size * 0.3 ) "
[FAIL] CSS Animations: property <min-height> from neutral to [min-content] at (0.6) should be [calc-size(min-content, 4px + size * 0.6)]
  assert_equals: expected "calc - size ( min - content , 4px + size * 0.6 ) " but got "calc - size ( min - content , 4px + 0px + size * 0.6 ) "
[FAIL] CSS Animations: property <min-height> from neutral to [min-content] at (1) should be [calc-size(min-content, 0px + size * 1)]
  assert_equals: expected "calc - size ( min - content , 0px + size * 1 ) " but got "calc - size ( min - content , 0px + 0px + size * 1 ) "
[FAIL] CSS Animations: property <min-height> from neutral to [min-content] at (1.5) should be [calc-size(min-content, -5px + size * 1.5)]
  assert_equals: expected "calc - size ( min - content , - 5px + size * 1.5 ) " but got "calc - size ( min - content , - 5px + 0px + size * 1.5 ) "
[FAIL] Web Animations: property <min-height> from neutral to [min-content] at (-0.3) should be [calc-size(min-content, 13px + size * -0.3)]
  assert_equals: expected "calc - size ( min - content , 13px + size * - 0.3 ) " but got "calc - size ( min - content , 13px + 0px + size * - 0.3 ) "
[FAIL] Web Animations: property <min-height> from neutral to [min-content] at (0) should be [calc-size(min-content, 10px + size * 0)]
  assert_equals: expected "calc - size ( min - content , 10px + size * 0 ) " but got "calc - size ( min - content , 10px + 0px + size * 0 ) "
[FAIL] Web Animations: property <min-height> from neutral to [min-content] at (0.3) should be [calc-size(min-content, 7px + size * 0.3)]
  assert_equals: expected "calc - size ( min - content , 7px + size * 0.3 ) " but got "calc - size ( min - content , 7px + 0px + size * 0.3 ) "
[FAIL] Web Animations: property <min-height> from neutral to [min-content] at (0.6) should be [calc-size(min-content, 4px + size * 0.6)]
  assert_equals: expected "calc - size ( min - content , 4px + size * 0.6 ) " but got "calc - size ( min - content , 4px + 0px + size * 0.6 ) "
[FAIL] Web Animations: property <min-height> from neutral to [min-content] at (1) should be [calc-size(min-content, 0px + size * 1)]
  assert_equals: expected "calc - size ( min - content , 0px + size * 1 ) " but got "calc - size ( min - content , 0px + 0px + size * 1 ) "
[FAIL] Web Animations: property <min-height> from neutral to [min-content] at (1.5) should be [calc-size(min-content, -5px + size * 1.5)]
  assert_equals: expected "calc - size ( min - content , - 5px + size * 1.5 ) " but got "calc - size ( min - content , - 5px + 0px + size * 1.5 ) "
Harness: the test ran to completion.
